generic-be/app/requirements.txt
pyyaml
pymongo
fastapi[all]
requests
uvicorn

isort
black